Extract mounted image asset provisioning from workspace prep
