+++ src/Foundry/Services/WinPe/IWinPeMountedImageAssetProvisioningService.cs
@@ -1,0 +1,10 @@
+namespace Foundry.Services.WinPe;
+
+internal interface IWinPeMountedImageAssetProvisioningService
+{
+    Task<WinPeResult> ProvisionAsync(
+        string mountedImagePath,
+        WinPeArchitecture architecture,
+        string? expertDeployConfigurationJson,
+        CancellationToken cancellationToken);
+}
