+++ src/Foundry/Services/WinPe/WinPeWorkspacePreparationService.cs
@@ -1,5 +1,4 @@
 using System.Globalization;
-using System.Text;
 using Microsoft.Extensions.Logging;
 
 namespace Foundry.Services.WinPe;
@@ -10,6 +9,7 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
     private readonly IWinPeDriverInjectionService _driverInjectionService;
     private readonly WinPeDriverPackageService _driverPackageService;
     private readonly IWinPeLocalDeployEmbeddingService _localDeployEmbeddingService;
+    private readonly IWinPeMountedImageAssetProvisioningService _mountedImageAssetProvisioningService;
     private readonly WinPeToolResolver _toolResolver;
     private readonly WinPeProcessRunner _processRunner;
     private readonly ILogger<WinPeWorkspacePreparationService> _logger;
@@ -19,6 +19,7 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
         IWinPeDriverInjectionService driverInjectionService,
         WinPeDriverPackageService driverPackageService,
         IWinPeLocalDeployEmbeddingService localDeployEmbeddingService,
+        IWinPeMountedImageAssetProvisioningService mountedImageAssetProvisioningService,
         WinPeToolResolver toolResolver,
         WinPeProcessRunner processRunner,
         ILogger<WinPeWorkspacePreparationService> logger)
@@ -27,6 +28,7 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
         _driverInjectionService = driverInjectionService;
         _driverPackageService = driverPackageService;
         _localDeployEmbeddingService = localDeployEmbeddingService;
+        _mountedImageAssetProvisioningService = mountedImageAssetProvisioningService;
         _toolResolver = toolResolver;
         _processRunner = processRunner;
         _logger = logger;
@@ -294,29 +296,6 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
         }
 
         _logger.LogInformation("Applied WinPE international settings successfully. MountDirectoryPath={MountDirectoryPath}", session.MountDirectoryPath);
-        string system32 = Path.Combine(session.MountDirectoryPath, "Windows", "System32");
-        Directory.CreateDirectory(system32);
-        string bootstrapScriptContent;
-        try
-        {
-            bootstrapScriptContent = WinPeDefaults.GetDefaultBootstrapScriptContent();
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to load embedded WinPE bootstrap script content.");
-            await session.DiscardAsync(cancellationToken).ConfigureAwait(false);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.InternalError,
-                "Failed to load embedded WinPE bootstrap script.",
-                ex.ToString());
-        }
-
-        await File.WriteAllTextAsync(
-            Path.Combine(system32, WinPeDefaults.DefaultBootstrapScriptFileName),
-            bootstrapScriptContent,
-            new UTF8Encoding(false),
-            cancellationToken).ConfigureAwait(false);
-        _logger.LogInformation("Wrote bootstrap script into mounted WinPE image. System32Path={System32Path}", system32);
 
         WinPeResult localDeployProvisioning = await _localDeployEmbeddingService.ProvisionAsync(
             session.MountDirectoryPath,
@@ -330,41 +309,16 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
         }
 
         _logger.LogInformation("Provisioned local Foundry.Deploy archive into mounted WinPE image. MountDirectoryPath={MountDirectoryPath}", session.MountDirectoryPath);
-        WinPeResult sevenZipProvisioning = ProvisionBundledSevenZipInImage(
+        WinPeResult assetProvisioning = await _mountedImageAssetProvisioningService.ProvisionAsync(
             session.MountDirectoryPath,
-            artifact.Architecture);
-        if (!sevenZipProvisioning.IsSuccess)
-        {
-            await session.DiscardAsync(cancellationToken).ConfigureAwait(false);
-            return sevenZipProvisioning;
-        }
-
-        _logger.LogInformation("Provisioned bundled 7-Zip tools into mounted WinPE image. MountDirectoryPath={MountDirectoryPath}", session.MountDirectoryPath);
-        WinPeResult deployConfigurationProvisioning = await ProvisionDeployConfigurationInImageAsync(
-            session.MountDirectoryPath,
+            artifact.Architecture,
             expertDeployConfigurationJson,
             cancellationToken).ConfigureAwait(false);
-        if (!deployConfigurationProvisioning.IsSuccess)
+        if (!assetProvisioning.IsSuccess)
         {
             await session.DiscardAsync(cancellationToken).ConfigureAwait(false);
-            return deployConfigurationProvisioning;
+            return assetProvisioning;
         }
-
-        if (!string.IsNullOrWhiteSpace(expertDeployConfigurationJson))
-        {
-            _logger.LogInformation("Provisioned Foundry.Deploy expert configuration into mounted WinPE image. MountDirectoryPath={MountDirectoryPath}", session.MountDirectoryPath);
-        }
-
-        string startnet = Path.Combine(session.MountDirectoryPath, WinPeDefaults.DefaultStartnetPathInImage);
-        string[] lines = File.Exists(startnet) ? await File.ReadAllLinesAsync(startnet, cancellationToken).ConfigureAwait(false) : ["wpeinit"];
-        var merged = lines.ToList();
-        if (!merged.Any(line => line.Contains(WinPeDefaults.DefaultBootstrapScriptFileName, StringComparison.OrdinalIgnoreCase)))
-        {
-            merged.Add(WinPeDefaults.DefaultBootstrapInvocation);
-        }
-
-        await File.WriteAllLinesAsync(startnet, merged, cancellationToken).ConfigureAwait(false);
-        _logger.LogInformation("Updated startnet.cmd in mounted WinPE image. StartnetPath={StartnetPath}", startnet);
 
         _logger.LogInformation("Committing mounted WinPE image changes. MountDirectoryPath={MountDirectoryPath}", session.MountDirectoryPath);
         WinPeResult commit = await session.CommitAsync(cancellationToken).ConfigureAwait(false);
@@ -516,126 +470,6 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
                 "Required WinPE optional components were not found in ADK.",
                 $"No required component CAB was found under '{ocRoot}'.");
     }
-    private WinPeResult ProvisionBundledSevenZipInImage(
-        string mountedImagePath,
-        WinPeArchitecture architecture)
-    {
-        string sourceRootPath = Path.Combine(AppContext.BaseDirectory, WinPeDefaults.BundledSevenZipRelativePath);
-        if (!Directory.Exists(sourceRootPath))
-        {
-            _logger.LogWarning("Bundled 7-Zip assets folder not found. SourceRootPath={SourceRootPath}", sourceRootPath);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.ToolNotFound,
-                "Bundled 7-Zip assets were not found.",
-                $"Expected path: '{sourceRootPath}'.");
-        }
-
-        string runtimeFolder = architecture.ToSevenZipRuntimeFolder();
-        string sourceExecutablePath = Path.Combine(sourceRootPath, runtimeFolder, "7za.exe");
-        if (!File.Exists(sourceExecutablePath))
-        {
-            _logger.LogWarning("Bundled 7-Zip executable not found for runtime folder {RuntimeFolder}. Path={ExecutablePath}", runtimeFolder, sourceExecutablePath);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.ToolNotFound,
-                "Bundled 7-Zip executable was not found for target architecture.",
-                $"Expected file: '{sourceExecutablePath}'.");
-        }
-
-        string sourceLicensePath = Path.Combine(sourceRootPath, "License.txt");
-        if (!File.Exists(sourceLicensePath))
-        {
-            _logger.LogWarning("Bundled 7-Zip license file not found. Path={LicensePath}", sourceLicensePath);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.ToolNotFound,
-                "Bundled 7-Zip license file was not found.",
-                $"Expected file: '{sourceLicensePath}'.");
-        }
-
-        string sourceReadmePath = Path.Combine(sourceRootPath, "readme.txt");
-        if (!File.Exists(sourceReadmePath))
-        {
-            _logger.LogWarning("Bundled 7-Zip readme file not found. Path={ReadmePath}", sourceReadmePath);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.ToolNotFound,
-                "Bundled 7-Zip readme file was not found.",
-                $"Expected file: '{sourceReadmePath}'.");
-        }
-
-        string destinationExecutablePath = Path.Combine(
-            mountedImagePath,
-            WinPeDefaults.EmbeddedSevenZipToolsPathInImage,
-            runtimeFolder,
-            "7za.exe");
-        string destinationToolsRootPath = Path.Combine(mountedImagePath, WinPeDefaults.EmbeddedSevenZipToolsPathInImage);
-
-        string? destinationDirectoryPath = Path.GetDirectoryName(destinationExecutablePath);
-        if (string.IsNullOrWhiteSpace(destinationDirectoryPath))
-        {
-            return WinPeResult.Failure(
-                WinPeErrorCodes.InternalError,
-                "Failed to resolve destination path for bundled 7-Zip provisioning.",
-                $"Destination file: '{destinationExecutablePath}'.");
-        }
-
-        try
-        {
-            Directory.CreateDirectory(destinationDirectoryPath);
-            Directory.CreateDirectory(destinationToolsRootPath);
-            File.Copy(sourceExecutablePath, destinationExecutablePath, overwrite: true);
-            File.Copy(sourceLicensePath, Path.Combine(destinationToolsRootPath, "License.txt"), overwrite: true);
-            File.Copy(sourceReadmePath, Path.Combine(destinationToolsRootPath, "readme.txt"), overwrite: true);
-            return WinPeResult.Success();
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to provision bundled 7-Zip tools into mounted WinPE image. DestinationToolsRootPath={DestinationToolsRootPath}", destinationToolsRootPath);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.BuildFailed,
-                "Failed to provision bundled 7-Zip executable into mounted WinPE image.",
-                ex.ToString());
-        }
-    }
-
-    private async Task<WinPeResult> ProvisionDeployConfigurationInImageAsync(
-        string mountedImagePath,
-        string? expertDeployConfigurationJson,
-        CancellationToken cancellationToken)
-    {
-        if (string.IsNullOrWhiteSpace(expertDeployConfigurationJson))
-        {
-            return WinPeResult.Success();
-        }
-
-        string destinationPath = Path.Combine(mountedImagePath, WinPeDefaults.EmbeddedDeployConfigPathInImage);
-        string? destinationDirectoryPath = Path.GetDirectoryName(destinationPath);
-        if (string.IsNullOrWhiteSpace(destinationDirectoryPath))
-        {
-            return WinPeResult.Failure(
-                WinPeErrorCodes.InternalError,
-                "Failed to resolve destination path for Foundry.Deploy expert configuration.",
-                $"Destination file: '{destinationPath}'.");
-        }
-
-        try
-        {
-            Directory.CreateDirectory(destinationDirectoryPath);
-            await File.WriteAllTextAsync(
-                destinationPath,
-                expertDeployConfigurationJson,
-                new UTF8Encoding(false),
-                cancellationToken).ConfigureAwait(false);
-            return WinPeResult.Success();
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to provision Foundry.Deploy expert configuration into mounted WinPE image. DestinationPath={DestinationPath}", destinationPath);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.BuildFailed,
-                "Failed to provision Foundry.Deploy expert configuration into mounted WinPE image.",
-                ex.ToString());
-        }
-    }
-
     private static string GetOptionalComponentsRootPath(string kitsRootPath, WinPeArchitecture architecture)
     {
         return Path.Combine(
